pyproject.toml
[project]
name = "manage-breast-screening"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["django (>=5.1.7,<6.0.0)"]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-django = "^4.10.0"
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "manage_breast_screening.config.settings"
python_files = "tests.py test_*.py *_tests.py"
